+++ changelog.md
@@ -126,3 +126,5 @@
     to avoid fetching duplicated messages.
   * Add 'random' and 'hash' partitioner for produce APIs
   * Allow `brod_group_subscrber:assign_partitions/3` to return an updated cb_state
+  * `brod_client:get_connection` is back
+  * make it possible to run tests on both mac and linux